Free Roam and Static Camera Access
Pass `PXLNAV_OPTIONS.staticCamera = true/false` to initially load your Room as locked to camera positions or allow the player to freely walk around as they with.

Note, as with before, if there are no colliders, it will default to Static Cam and fail silently, as it expects if you want colliders, you'll add them into the FBX file to load in.

Example added, simple open area with a barrel.  It lets you run around and walk up uneven terrain and try out the navigation.

Added movementScalar, jumpScalar, & userScale added to pxlCamera to multiply the movement speed and jump impulse speed from outside of pxlNav.